--- types/emitter.ts
@@ -1,8 +1,0 @@
-export interface Emitter<Event = string> {
-  on(event: Event, listener: Function): Emitter<Event>;
-  once(event: Event, listener: Function): Emitter<Event>;
-  off(event?: Event, listener?: Function): Emitter<Event>;
-  emit(event: Event, ...args: any[]): Emitter<Event>;
-  listeners(event: Event): Function[];
-  hasListeners(event: Event): boolean;
-}
