+++ types/lib/apple-pay.ts
